Change ISBN type for long in controller, service and repository

+++ Library.API/Controllers/BookController.cs
@@ -18,7 +18,7 @@ namespace Library.API.Controllers
         }
 
         [HttpGet("ISBN")]
-        public async Task<BookViewModel> GetByIsbn([FromQuery(Name = "ISBN")] int isbn, CancellationToken cancellationToken)
+        public async Task<BookViewModel> GetByIsbn([FromQuery(Name = "ISBN")] long isbn, CancellationToken cancellationToken)
         {
             var book = await _service.GetByISBNAsync(isbn, cancellationToken);
